refactor: move url utilities to a shared path
